requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
click==8.2.2
fastapi==0.116.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
idna==3.10
loguru==0.7.3
psycopg==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.42
starlette==0.47.2
starlette-graphene3==0.6.0
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
websockets==15.0.1
